WIP: adding linkService and feed autodiscovery
- adding a link service that deals with adding/removing link entries in
the DOM (in the <head>) - adding feed links in the head in the main
syndicated video listings

BUG: the link removal is NOT WORKING correctly. Debug shows it finding
and removing entries, but the numbers are wrong (5 entries in the DOM,
and 3 in the output HTML). We are looking to keep the 3 per page. Yet,
they don't get removed from the autodiscovery list in Firefox… Thus,
this is more 2-3 bugs that might very well be related. Send halp.